# Conflicts:
#	app/src/main/java/gr/aueb/carpooling/model/Request_status.java
#	app/src/main/java/gr/aueb/carpooling/model/view/passenger/ExistedSubroutes/ExistedSubrouteActivity.java

+++ app/src/main/java/gr/aueb/carpooling/model/view/passenger/ExistedSubroutes/ExistedSubrouteActivity.java
@@ -6,12 +6,13 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.app.AlertDialog;
-
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 
+import java.text.DecimalFormat;
 
 import gr.aueb.carpooling.R;
 import gr.aueb.carpooling.model.Passenger;
@@ -25,10 +26,10 @@ import gr.aueb.carpooling.model.dao.SubrouteDAO;
 import gr.aueb.carpooling.model.memoryDao.PassengerDAOmemory;
 import gr.aueb.carpooling.model.memoryDao.RouteDAOmemory;
 import gr.aueb.carpooling.model.memoryDao.SubrouteDAOmemory;
-
+import gr.aueb.carpooling.model.view.driver.createRoute.CreateRouteActivity;
 import gr.aueb.carpooling.model.view.passenger.DriverRating.DriverRatingActivity;
 import gr.aueb.carpooling.model.view.passenger.front_page.PassengerFrontPageActivity;
-
+import gr.aueb.carpooling.model.view.passenger.top_up.TopUpActivity;
 
 public class ExistedSubrouteActivity extends AppCompatActivity implements ExistedSubrouteView, ExistedSubrouteRecyclerViewAdapter.SubrouteSelectionListener {
 
@@ -72,8 +73,8 @@ public class ExistedSubrouteActivity extends AppCompatActivity implements Existe
 
     public void selectSubroute(Subroute subroute, Request_status status,Boolean b) {
         //showErrorMessage("Dest",subroute.getDestination().toString());
-        Money cost=subroute.calculateCost();
-        //showErrorMessage("cost",String.valueOf(cost.getAmount()));
+//        Money cost=subroute.calculateCost();
+//        showErrorMessage("cost",String.valueOf(cost.getAmount()));
 //        Boolean success= passenger.transaction(cost);
 //        if(!success){
 //            new androidx.appcompat.app.AlertDialog.Builder(this)
@@ -92,11 +93,19 @@ public class ExistedSubrouteActivity extends AppCompatActivity implements Existe
 
         if(b){
             if (status == Request_status.APPROVED){
-                subroute.setStatus(Request_status.COMPLETED);
-                Intent intent = new Intent(ExistedSubrouteActivity.this, DriverRatingActivity.class);
-                intent.putExtra("Username", username);
-                intent.putExtra("SubrouteId", subroute.getId());
-                startActivity(intent);
+                boolean success=payment(subroute);
+                if(success){
+                    subroute.setStatus(Request_status.COMPLETED);
+                    Intent intent = new Intent(ExistedSubrouteActivity.this, DriverRatingActivity.class);
+                    intent.putExtra("Username", username);
+                    intent.putExtra("SubrouteId", subroute.getId());
+                    startActivity(intent);
+                }else{
+                    Intent intent = new Intent(ExistedSubrouteActivity.this,TopUpActivity.class);
+                    intent.putExtra("Username", username);
+                    startActivity(intent);
+                }
+
             } else if (status == Request_status.REJECTED) {
                 showErrorMessage("You can press button complete if request status is approved. Current status ", String.valueOf(Request_status.REJECTED));
             } else if(status == Request_status.PENDING){
@@ -105,13 +114,18 @@ public class ExistedSubrouteActivity extends AppCompatActivity implements Existe
                 showErrorMessage("You have already completed this subroute. Current status: ", String.valueOf(Request_status.COMPLETED));
             }
         } else {
+            if (status == Request_status.APPROVED) {
+                showErrorMessage("You cann't press button delete if request status is approved.Now is: ", String.valueOf(Request_status.APPROVED));
+            }else{
+                    Route route = routeDAO.findRouteBySubroute(subroute);
+                    route.removeSubroute(subroute);
+                    subrouteDAO.delete(subroute);
+                    Intent intent = new Intent(this, this.getClass());
+                    intent.putExtra("Username", username);
+                    startActivity(intent);
+            }
 
-            Route route = routeDAO.findRouteBySubroute(subroute);
-            route.removeSubroute(subroute);
-            subrouteDAO.delete(subroute);
-            Intent intent = new Intent(this, this.getClass());
-            intent.putExtra("Username", username);
-            startActivity(intent);
+
         }
 
     }
@@ -149,6 +163,19 @@ public class ExistedSubrouteActivity extends AppCompatActivity implements Existe
 ////        intent.putExtra("Username",username);
 //        startActivity(intent);
 
+
+    }
+
+    @Override
+    public boolean payment(Subroute subroute) {
+        showErrorMessage("Balance",new DecimalFormat("0.00").format(passenger.getBalance()));
+        Money cost=subroute.calculateCost();
+        boolean success= passenger.transaction(cost);
+        if(!success){
+            showErrorMessage("Η πληρωμή επέτυχε.Πρόσθεσε χρήματα.Η διαδρομη κόστησε",new DecimalFormat("0.00").format(cost.getAmount()));
+        }
+
+        return success;
 
     }
